requirements.txt
python-dotenv
openai
flask_cors
flask
functions-framework>=3.4.0